@@ -1,4 +1,4 @@
-class CreateFlavors < ActiveRecord::Migration
+class CreateFlavors < ActiveRecord::Migration[7.0]
   def change
     create_table :flavors do |t|
       t.string :name
