@@ -1,4 +1,4 @@
-class DeviseCreateModels < ActiveRecord::Migration
+class DeviseCreateModels < ActiveRecord::Migration[7.0]
   def change
     create_table :models do |t|
       ## Database authenticatable
