@@ -1,4 +1,4 @@
-class AddImageToFlavors < ActiveRecord::Migration
+class AddImageToFlavors < ActiveRecord::Migration[7.0]
   def change
     add_column :flavors, :availability, :string
     add_column :flavors, :flavorofweek, :string
