@@ -1,4 +1,4 @@
-class AddNameToUsers < ActiveRecord::Migration
+class AddNameToUsers < ActiveRecord::Migration[7.0]
   def change
     add_column :users, :fname, :string
     add_column :users, :lname, :string
