@@ -1,4 +1,4 @@
-class CreateLeads < ActiveRecord::Migration
+class CreateLeads < ActiveRecord::Migration[7.0]
   def change
     create_table :leads do |t|
       t.string :fname
