@@ -1,4 +1,4 @@
-class AddInStockStatusFlavors < ActiveRecord::Migration
+class AddInStockStatusFlavors < ActiveRecord::Migration[7.0]
   def change
     add_column :flavors, :instock, :string
   end
