@@ -7,9 +7,9 @@ class User < ApplicationRecord
   after_create :subscribe_user_to_list
 
   def subscribe_user_to_list
-    gibbon = Gibbon::Request.new(api_key: ENV['MAILCHIMP_API_KEY'], debug: true)
+    gibbon = Gibbon::Request.new(api_key: Rails.application.credentials.mailchimp_api_key, debug: true)
     begin
-      gibbon.lists(ENV["MAILCHIMP_LIST_ID_"]).members.create(body: { 
+      gibbon.lists(Rails.application.credentials.mailchimp_list_id).members.create(body: {
         email_address: self.email,
         status: "subscribed",
         merge_fields: {:FNAME => self.fname, :LNAME => self.lname}
