@@ -1,5 +1,5 @@
 # This migration comes from active_storage (originally 20190112182829)
-class AddServiceNameToActiveStorageBlobs < ActiveRecord::Migration[6.0]
+class AddServiceNameToActiveStorageBlobs < ActiveRecord::Migration[7.0]
   def up
     return unless table_exists?(:active_storage_blobs)
 
